feat(scripts): module energy collecting made more dynamic

+++ scripts/run_wfey.sh
@@ -4,7 +4,7 @@
 
 source ./scripts/env.sh
 set -m # enables job control
-#set -x 
+set -x 
 #VERBOSE=1
 
 SCRIPTDIR=$(dirname $0)
@@ -18,7 +18,8 @@ if [ ! -f $WFEYCONFIG ]; then make; fi
 
 ARGSTR=${PARAMETERS// /_}
 
-LOGDATE=$(date +%Y-%m-%d-%H-%M-%S)
+#LOGDATE=$(date +%Y-%m-%d-%H-%M-%S)
+echo $LOGDATE
 
 LOGPATH="logs/$LOGS/$WFEYCONFIG/$ARGSTR/"
 
@@ -37,25 +38,29 @@ fi
 mkdir -p $LOGPATH
 
 
-# --------- Run ENERGY SCRIPT --------- #
-## This script should contain a function called 'powerLog' which should
-## take in the Path and the Data args to create a file '$LOGPATH/hwmon-${LOGDATE}.out'
+if [[ -n "$BM_ENERGY" ]]; then
+    # --------- Run ENERGY SCRIPT --------- #
+    ## This script should contain a function called 'powerLog' which should
+    ## take in the Path and the Data args to create a file '$LOGPATH/hwmon-${LOGDATE}.out'
 
-## NOTE -- I believe due to inheritance, energyscript should also be
-## taskset to the runnercpu core but I have not confirmed this yet
-source $ENERGYSCRIPT
+    ## NOTE -- I believe due to inheritance, energyscript should also be
+    ## taskset to the runnercpu core but I have not confirmed this yet
+    source $ENERGYSCRIPT
 
-powerLog "$LOGPATH" "$LOGDATE" &
-powerPID=$!
+    powerLog "$LOGPATH" "$LOGDATE" &
+    powerPID=$!
+fi
 
 # ---------- Run WFEY ---------- #
 
 ./$@ 1> $LOGPATH/latency-${LOGDATE}.out 2> $LOGPATH/wfey-${LOGDATE}.out
 
-sleep ${SLEEP_TO_FINISH} # Power numbers from hwmon continue to go up a little after the wfey code is done
-
-if [[ "$ENERGYTYPE"=="NOMOD" ]]; then
-    kill -USR1 $powerPID
-elif [[ "$ENERGYTYPE"=="WMOD" ]]
-    cleanup "$LOGPATH" "$LOGDATE"
+if [[ -n "$BM_ENERGY" ]]; then
+    sleep ${SLEEP_TO_FINISH} # Power numbers from hwmon continue to go up a little after the wfey code is done
+    
+    if [[ "$ENERGYTYPE"=="NOMOD" ]]; then
+	kill -USR1 $powerPID
+    elif [[ "$ENERGYTYPE"=="WMOD" ]]
+	 cleanup "$LOGPATH" "$LOGDATE"
+    fi
 fi
